==> exploracion/01-apertura_archivos.ipynb <==
import pandas as pd
import numpy as np
import os 
import xml.etree.ElementTree as ET


os.chdir("data")

os.getcwd()

df_csv = pd.read_csv("01-data_remaning_kaggle.csv")
df_csv.head(2)

df_exported_data1= pd.read_csv("01-exported_db_data.csv")
df_exported_data1.head(2)

data_xml = pd.read_xml("01-data_xml_clean.xml")
data_xml.head()
        
      

df_exported_txt= pd.read_csv("01-data_txt_clean.txt", sep= ';')
df_exported_txt.head(2)

df_csv.to_csv("02-data_remaning_kaggle.csv")

df_exported_data1.to_csv("02-exported_db_data.csv")

data_xml.to_csv("02-data_xml_clean.csv")


df_exported_txt.to_csv("02-data_txt_clean.csv")
