==> exploracion/02-eliminar-info.ipynb <==
import pandas as pd
import numpy as np
import os 
import xml.etree.ElementTree as ET


pd.options.display.max_columns = None

df_csv = pd.read_csv("data/01-data_remaning_kaggle.csv")

df_csv.head(2)

df_csv.columns

columnas_eliminadas_q18 = df_csv.drop(["Q18_Part_1", "Q18_Part_2",	"Q18_Part_3",	"Q18_Part_4",	"Q18_Part_5",	"Q18_Part_6",	"Q18_OTHER"], axis = 1, inplace = True)

df_csv.columns

columnas_eliminadas_q19 = df_csv.drop(['Q19_Part_1', 'Q19_Part_2', 'Q19_Part_3','Q19_Part_4', 'Q19_Part_5', 'Q19_OTHER'], axis = 1, inplace = True)

columnas_eliminadas_q28 = df_csv.drop(["Q28"], axis = 1, inplace = True)

df_csv.columns

columnas_eliminadas_q31 = df_csv.drop(["Q31_A_Part_1","Q31_A_Part_2","Q31_A_Part_3","Q31_A_Part_4","Q31_A_Part_5","Q31_A_Part_6","Q31_A_Part_7","Q31_A_Part_8","Q31_A_Part_9","Q31_A_OTHER"], axis = 1, inplace = True)

columnas_eliminadas_q36 = df_csv.drop(["Q36_A_Part_1","Q36_A_Part_2","Q36_A_Part_3","Q36_A_Part_4","Q36_A_Part_5","Q36_A_Part_6","Q36_A_Part_7","Q36_A_OTHER"], axis = 1, inplace = True)

columnas_eliminadas_q3738 = df_csv.drop(["Q37_A_Part_1"	,"Q37_A_Part_2","Q37_A_Part_3", "Q37_A_Part_4", "Q37_A_Part_5", "Q37_A_Part_6", "Q37_A_Part_7", "Q37_A_OTHER", "Q38_A_Part_1",	"Q38_A_Part_2",	"Q38_A_Part_3", "Q38_A_Part_4", "Q38_A_Part_5", "Q38_A_Part_6",	"Q38_A_Part_7", "Q38_A_Part_8", "Q38_A_Part_9", "Q38_A_Part_10", "Q38_A_Part_11","Q38_A_OTHER"], axis = 1, inplace = True)

# columnas_eliminadas_q3738 = df_csv.drop(["Q37_A_Part_1"	,"Q37_A_Part_2","Q37_A_Part_3", "Q37_A_Part_4", "Q37_A_Part_5", "Q37_A_Part_6", "Q37_A_Part_7", "Q37_A_OTHER", "Q38_A_Part_1",	"Q38_A_Part_2",	"Q38_A_Part_3", "Q38_A_Part_4", "Q38_A_Part_5", "Q38_A_Part_6",	"Q38_A_Part_7", "Q38_A_Part_8", "Q38_A_Part_9", "Q38_A_Part_10", "Q38_A_Part_11","Q38_A_OTHER"], axis = 1, inplace = True)

df_csv.head(2)

df_txt = pd.read_csv("data/02-data_txt_clean.csv")

df_txt.head(2)

df_txt.columns

columnas_txt_eli = df_txt.drop(['Unnamed: 0','Q15', 'Q16', 'Q23','Q17','Q34'], axis = 1, inplace = True)

columnas_txt_el =df_txt.drop(['Q35'], axis = 1, inplace = True)

df_txt.columns
df_txt.head(2)

os.getcwd()

df_txt.to_csv("data/03-data_txt_eliminados.csv")

df_csv.to_csv("data/03-data_csv_eliminados.csv")

df_xml = pd.read_csv("data/02-data_xml_clean.csv")

df_xml.columns

columnas_xml_eli = df_xml.drop(['Unnamed: 0','level_0'], axis = 1, inplace = True)

df_xml.head()

df_xml.to_csv("data/03-data_xml_eliminados.csv")

df_db = pd.read_csv("data/02-exported_db_data.csv")

df_db.columns

columnas_db_eli = df_db.drop(['Unnamed: 0'], axis = 1, inplace = True)

df_db.to_csv("data/03-data_db_eliminados.csv")
