==> exploracion/03-renombrando-columnas.ipynb <==
import pandas as pd


pd.options.display.max_columns = None

df_txt= pd.read_csv("data/03-data_txt_eliminados.csv", )
df_txt.head(2)

df_txt.rename(columns = {"Q3": "pais", "Q4": "educacion", "Q5": "ocupacion", "Q6":  "experiencia", "Q7": "l_utilizado", "Q8":"l_recomendado", "Q9": "adi", "Q11":"plataforma", "Q12": "hardware", "Q13": "tpu", "Q14":"visualizacion", "Q20": "i_actual", "Q21": "e_tamanio", "Q22": "e_analistas", "Q24": "e_actividades", "Q25": "salario_actual", "Q26": "presupuesto_ml_cc", "Q32": "big_products" , "Q33": "big_products_used" , "Q41": "tool_ppal" }, inplace = True)
df_txt.head(2)

df_data= pd.read_csv("data/03-data_db_eliminados.csv")
df_data.head(5)

df_data.rename(columns = {"Q10_Part_1": "kaggle_n","Q10_Part_2":"colab_n", "Q10_Part_3": "azure_n","Q10_Part_4": "paper_space_gradient" ,"Q10_Part_5": "binder_jupyter_hub", "Q10_Part_6" : "code_ocean", "Q10_Part_7" : "ibm_ws", "Q10_Part_8" : "amazon_ssn",	"Q10_Part_9" : "amazoner_n", "Q10_Part_10": "googlec_n", "Q10_Part_11" : "googlec_d", "Q10_Part_12" : "databricksc_n", "Q10_Part_13" : "zeppelin_zn", "Q10_Part_14" : "deepnote_n",	"Q10_Part_15" : "obs_n", "Q10_Part_16": "none", "Q10_OTHER" : "other"}, inplace = True)
df_data.head(5)

df_txt.to_csv("data/04-data-txt-col-rename.csv")

df_data.to_csv("data/04-data-db-col-rename.csv")
