==> exploracion/05-union.ipynb <==
import pandas as pd
import numpy as np
import os
pd.options.display.max_columns = None

# db 

df_db= pd.read_csv("data/01-exported_db_data.csv")
df_db.head(2)

# txt

df_txt= pd.read_csv("data/01-data_txt_clean.txt",sep= ';')
df_txt.head(2)

# xml

df_xml= pd.read_xml("data/01-data_xml_clean.xml")
df_xml.head(2)

# kaggle

df_csv= pd.read_csv("data/01-data_remaning_kaggle.csv",index_col=0 )
df_csv.head(2)

df_csv.columns

# Unirlos

# txt+xml

'''df_xml_txt = df_xml.merge(df_txt, on='index')
df_xml_txt.head(2)'''

# xml(genero) + txt +db(Q10)

'''df_xml_txt_db = df_xml_txt.merge(df_db, on='index')
df_xml_txt_db.head(2)'''

# xml(genero) + txt +db(Q10) + kaggle ( el resto separadas)

'''df_xml_txt_db_csv = df_xml_txt_db.merge(df_csv, on='index')
df_xml_txt_db_csv.head(2)'''

#guardarlo en un archivo
#df_xml_txt_db_csv.to_csv("data/05-data_union.csv")


def union(df_txt,df_xml,df_db,df_csv):#db,kagle, txt. xml
    df_txt.set_index("index", inplace =True)
    df_xml.set_index("index", inplace =True)
    df_db.set_index("index", inplace =True)
    df_csv.set_index("index", inplace =True)

    u_1= df_csv.join(df_db, on = "index")#txt + xml gender 
    u_2=u_1.join(df_txt, on = "index")#db Q10
    u_fin=u_2.join(df_xml, on = "index")#kaggle Q18
    return u_fin


df = union(df_db,df_csv,df_txt,df_xml)

df.head(2)

#guardarlo en un archivo
df.to_csv("data/05-data_union.csv")
