==> exploracion/06-eliminacion_extra.ipynb <==
import pandas as pd
import numpy as np
import os 
import xml.etree.ElementTree as ET
pd.options.display.max_columns = None

df_unido = pd.read_csv('data/05-data_union.csv', index_col = 0)

df_unido.head(1)

df_unido.columns

columnas_extra_eliminadas1 = df_unido.drop( ['tpu','hardware',],axis = 1 )

q29a_eliminada = df_unido.drop(['Q29_A_Part_1','Q29_A_Part_2','Q29_A_Part_3','Q29_A_Part_4'], axis = 1)

q30a_eliminada = df_unido.drop(['Q30_A_Part_1', 'Q30_A_Part_2',	'Q30_A_Part_3',	'Q30_A_Part_4',	'Q30_A_Part_5',	'Q30_A_Part_6', 'Q30_A_Part_7', 'Q30_A_OTHER'], axis = 1)

#de aqui en adelante aprecen estar ya elimninadas, chequear 
# q31a_eliminada = df_unido.drop([''])
#si estamos de acuerdo correr y guardar el archivo un poco mas limpio
