==> exploracion/07-limpio.ipynb <==
import pandas as pd
import numpy as np
import os 
import xml.etree.ElementTree as ET
pd.options.display.max_columns = None
import re

df = pd.read_csv('data/05-data_union.csv', index_col = 0)

df_eliminar = df.drop(['tpu','hardware', 'e_actividades','Q29_A_Part_1', 'Q29_A_Part_2', 'Q29_A_Part_3', 'Q29_A_Part_4','Q29_A_OTHER', 'Q30_A_Part_1', 'Q30_A_Part_2', 'Q30_A_Part_3', 'Q30_A_Part_4', 'Q30_A_Part_5', 'Q30_A_Part_6', 'Q30_A_Part_7', 'Q30_A_OTHER'], axis = 1)

# Función para dividir los valores de las columnas problemáticas en nuevas columnas

df_eliminar['big_products'].unique()

x = df_eliminar["big_product_lista"].tolist()
x[:2]

set1 = set()

for lista in x:
    try:
        for herramienta in lista:
            set1.add(herramienta.strip())
    except:
        pass
lista_columnas_nuevas = list(set1)
lista_columnas_nuevas

for i in lista_columnas_nuevas:
    df_eliminar[i] = df_eliminar['big_products'].str.contains(i, regex = True)

df_eliminar.head(2)

# def separar(col):
    # for col in lista_columnas_nuevas:
    # df_eliminar[i] = df_eliminar['big_products'].str.contains(i, regex = True)

df["Comentario"] = df["Comentario"].apply(lambda x: int(x)) #1 if str(x)=="True" else 0

import seaborn as sns

# sns.countplot(x= lista_columnas_nuevas , data= df_eliminar)

# fig = sns.countplot(data = df_eliminar, x = 'big_products', hue = 'Amazon Redshift', palette='Paired')

df_eliminar.to_csv("data/6-data-limpieza.csv")
