==> exploracion/08-funciones-columnas.ipynb <==
## Código para columnas con contenido múltiple en celdas

import pandas as pd
import numpy as np
import os 
import xml.etree.ElementTree as ET
import seaborn as sns
import re
import matplotlib.pyplot as plt
pd.options.display.max_columns = None

df = pd.read_csv('data/05-data_union.csv', index_col = 0)

df.head(1)

df2 = df.drop(['tpu','hardware', 'e_actividades','Q29_A_Part_1', 'Q29_A_Part_2', 'Q29_A_Part_3', 'Q29_A_Part_4','Q29_A_OTHER', 'Q30_A_Part_1', 'Q30_A_Part_2', 'Q30_A_Part_3', 'Q30_A_Part_4', 'Q30_A_Part_5', 'Q30_A_Part_6', 'Q30_A_Part_7', 'Q30_A_OTHER'], axis = 1)

Función para separar los datos múltiples

def separar(df,col,col2):
    df[col2] = df2[col].str.split(",")
    x = df[col2].tolist()

    set1 = set() #Creamos un set para que no salgan los contenidos repetidos

    for lista in x:
        try:
            for herramienta in lista:
                herramienta = herramienta.strip()
                set1.add(herramienta)
        except:
            pass
    lista_columnas_nuevas = list(set1)

    return lista_columnas_nuevas 


lista_lutilizados = separar(df2,'l_utilizado','l_utilizado2')

lista_lutilizados[:10]

for i in lista_lutilizados:
    df2[i] = df2['l_utilizado'].str.contains(i)

df2.head()

df2['l_utilizado2'] = df2['l_utilizado'].str.split(",")
x = df['l_utilizado'].tolist()

set1 = set() #Creamos un set para que no salgan los contenidos repetidos

for lista in x:
    try:
        for herramienta in lista:
            herramienta = herramienta.strip()
            set1.add(herramienta)
    except:
        pass
lista_columnas_nuevas2 = list(set1)

for i in lista_columnas_nuevas2:
    df2[i] = df2['l_utilizado'].str.contains(i)

## Función para generar las gráficas con las columnas de booleanos

def graficar(df,df_nuevo):

    df_nuevo = df.melt(value_vars = df.columns)
    df_nuevo = df_nuevo[df_nuevo['value']==True].reset_index()
    df_agrup = df_nuevo.groupby(['variable'])['value'].count().reset_index()
    df_ordenado = df_agrup.sort_values('value', ascending = False)
    df_def = df_ordenado.drop_duplicates('value', keep = 'last')

    return df_def


graficar()
